Move function which generated composite URLMap to translator pkg
